tests/wpt/web-platform-tests/app-history/navigate/state/updateCurrent/state-after-reload.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    i.contentWindow.appHistory.updateCurrent({ state: { data: "value" } });
    assert_equals(i.contentWindow.appHistory.entries().length, 1);
    assert_equals(i.contentWindow.appHistory.current.getState().data, "value");

    i.contentWindow.location.reload();
    i.onload = t.step_func_done(() => {
      assert_equals(i.contentWindow.appHistory.entries().length, 1);
      assert_equals(i.contentWindow.appHistory.current.getState().data, "value");
    });
  });
}, "appHistoryEntry.getState() after location.reload()");
</script>
